pyproject.toml
[tool.poetry]
name = "mediation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.4"
psycopg2-binary = "^2.9.6"
django-debug-toolbar = "^4.1.0"
pillow = "^10.0.0"
sorl-thumbnail = "^12.9.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
django-ckeditor = "^6.7.0"
singlemodeladmin = "^0.9"
django-reversion = "^5.0.4"
django-environ = "^0.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
